Beginnings of a "contacts" sample app.
So far, this has mappings to a subset of the contacts content-provider
schema, and a dummy activity which dumps enough out to the log to verify
that it's getting processed properly.

Dumped output includes a list of contacts, and for each contact, all
'data' records.

Phone and Email data rows are identified as such, and mapped to
appropriate subclasses of the generic ContactData base class; we also
figure out what 'subtype' of phone or email that we're looking at, and
dig the appropriate type string out of resources.

All other data records are mapped to a generic 'UnknownData' row type;
we dump out the row's "mimetype", which indicates what kind of row we're
looking at, and 'data1', which is, well... likely to be something interestin
(though it isn't for photos).