Mapping for StructuredName ContactData records.
This requires some new forms of special treatment; in particular, we
do *not* want to send any of the "structured" fields unless the user
meant to set them, as that keeps the NameSplitter in the content
provider from setting them itself.

Also, the beginning of tests for the app itself.