Slicker interface to startActivityForResult and onActivityResult.
Adds an "ActivityResultDispatch" trait for activities which allows
activity code to do

  withActivityResult( someIntent ){ (resultCode, resultIntent) =>
     ...
  }

This allocates a new "request code" (by bumping a counter), registers
the function in the second argument as the callback for results with
that request code, and then calls startActivityForResult as normally.
When onActivityResult is called (as the activity restarts), we look
for a handler for the request code we got, and invoke it if found.

It's sometimes useful to be able to have access to this functionality
not just in the activity itself, but also in associated widgets.
The "sample contacts" view utilities are augmented with code to do
that.  It works by calling getContext (and baseContext if it's a
wrapper) to find the underlying activity (which must have the
ActivityResultDispatch trait mixed in), and then delegating
"withActivityResult" to it.