Contacts app:  when aggregating ContactData, track what acct it came from.
There are cases where we need to know the account info.  For one thing,
we're using account-info objects to determine what string we use to
display a category label.  For another, we'll need the account info to
determine how to display "unknown" data items associated with, e.g.,
Facebook accounts.

So, the aggregator now spits out "AggregatedDatum" objects, which pair
up a ContactData with the AccountInfo of the account that it came from.