Initial data kind check code.
Intended to restrict the numbers of ContactData rows of particular "kind"s
to what an account of that type can accommodate.  This is mainly a way of
dealing with restrictions imposed by particular SyncAdapters (or, more
precisely, by the servers of whatever type that they're syncing with).
That, in turn, seems to be, in practice, mostly a way of accommodating
Exchange servers...